requirements.txt
absl-py==0.9.0
appdirs==1.4.3
astor==0.8.1
attrs==19.3.0
black==19.10b0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cycler==0.10.0
entrypoints==0.3
flake8==3.7.9
gast==0.2.2
google-auth==1.13.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
joblib==0.15.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.2.1
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.2
oauthlib==3.1.0
opt-einsum==3.2.0
pandas==1.0.3
pathspec==0.8.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pyflakes==2.1.1
pymongo==3.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2019.3
PyYAML==5.3.1
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.23.1
scipy==1.4.1
six==1.14.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==1.0.1
wrapt==1.12.1
